Add missing simpleeval tests for base_n, allocate only one SimpleEval for base_n
Create a utils.singleton.singleton function that acts as a decorator
- Add tests under tests.test_utils

Move the simpleeval.SimpleEval instantiation outside of BaseNQueryHandler
as a function decorated with singleton. This makes each BaseHandler to have
only one instance of SimpleEval instead of multiple.

Minor fixes in queries and hex results to avoid showing empty placeholders for no query
or first space converted to bytes.